Fix codegen unittests on Linux and various compiler warnings while building codegen.

+++ src/backend/codegen/op_expr_tree_generator.cc
@@ -92,7 +92,8 @@ bool OpExprTreeGenerator::VerifyAndCreateExprTree(
   List *arguments = reinterpret_cast<FuncExprState*>(expr_state)->args;
   assert(nullptr != arguments);
   // In ExecEvalFuncArgs
-  assert(list_length(arguments) == itr->second->GetTotalArgCount());
+  assert(list_length(arguments) ==
+        static_cast<int>(itr->second->GetTotalArgCount()));
 
   ListCell   *arg = nullptr;
   bool supported_tree = true;
